Move Metadata Interceptor tests to new directory
